refactor: split pre-task, simulation
